+++ jax/interpreters/partial_eval.py
@@ -26,7 +26,7 @@ from weakref import ref
 
 import numpy as np
 
-from jax import linear_util as lu
+from jax._src import linear_util as lu
 from jax.config import config
 from jax._src import api_util
 from jax._src import core
