+++ tests/jaxpr_effects_test.py
@@ -21,7 +21,7 @@ import jax
 import jax.numpy as jnp
 from jax import core
 from jax import lax
-from jax import linear_util as lu
+from jax._src import linear_util as lu
 from jax.config import config
 from jax.interpreters import ad
 from jax.experimental import maps
